Move adapter stuff to its own package
