+++ pipeline/inferrer/inference_manager/src/main/scala/uk/ac/wellcome/platform/inference_manager/adapters/FeatureVectorInferrerAdapter.scala
@@ -1,13 +1,9 @@
-package uk.ac.wellcome.platform.inference_manager.services
+package uk.ac.wellcome.platform.inference_manager.adapters
 
 import java.nio.{ByteBuffer, ByteOrder}
 import java.util.Base64
 
-import akka.http.scaladsl.model._
-import akka.http.scaladsl.unmarshalling.Unmarshal
-import akka.stream.Materializer
-import de.heikoseeberger.akkahttpcirce.FailFastCirceSupport._
-import grizzled.slf4j.Logging
+import akka.http.scaladsl.model.{HttpMethods, HttpRequest, Uri}
 import io.circe.Decoder
 import io.circe.generic.semiauto.deriveDecoder
 import uk.ac.wellcome.models.work.internal.{AugmentedImage, InferredData}
@@ -16,43 +12,7 @@ import uk.ac.wellcome.platform.inference_manager.models.{
   FeatureVectorInferrerResponse
 }
 
-import scala.concurrent.Future
 import scala.util.{Success, Try}
-
-/*
- * An InferrerAdapter is specific to the inferrer and the data that is being augmented.
- * Implementors must provide:
- * - The type of the inferrer response
- * - A Decoder for that response
- * - A function to create an HTTP request from the input data type
- * - A function to augment input data with a response, returning the output data type
- *
- * Additionally, the trait provides the logic for handling different HTTP response statuses
- */
-
-trait InferrerAdapter[Input, Output] extends Logging {
-  type InferrerResponse
-
-  implicit val responseDecoder: Decoder[InferrerResponse]
-  def createRequest(input: Input): HttpRequest
-  def augmentInput(input: Input,
-                   inferrerResponse: Option[InferrerResponse]): Output
-
-  def parseResponse(response: HttpResponse)(
-    implicit mat: Materializer): Future[Option[InferrerResponse]] =
-    response.status match {
-      case StatusCodes.OK =>
-        Unmarshal(response.entity).to[Some[InferrerResponse]]
-      case StatusCodes.BadRequest =>
-        Future.failed(new Exception("Bad request"))
-      case StatusCodes.NotFound =>
-        Future.failed(new Exception("Entity not found"))
-      case statusCode =>
-        warn(
-          s"Request failed non-deterministically with code ${statusCode.value}")
-        Future.successful(None)
-    }
-}
 
 // The InferrerAdaptor for feature vectors, consuming MergedImages and
 // augmenting them into AugmentedImages
