+++ pipeline/inferrer/inference_manager/src/test/scala/uk/ac/wellcome/platform/inference_manager/services/InferenceManagerWorkerServiceTest.scala
@@ -12,6 +12,7 @@ import uk.ac.wellcome.messaging.memory.MemoryMessageSender
 import uk.ac.wellcome.models.work.internal.{AugmentedImage, InferredData}
 import uk.ac.wellcome.models.Implicits._
 import uk.ac.wellcome.models.work.generators.ImageGenerators
+import uk.ac.wellcome.platform.inference_manager.adapters.FeatureVectorInferrerAdapter
 import uk.ac.wellcome.platform.inference_manager.fixtures.{
   InferenceManagerWorkerServiceFixture,
   MemoryFileWriter,
